Moved deleting invalid projects to IBuildSupport
