+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/managers/GradleBuildSupport.java
@@ -14,10 +14,13 @@ package org.eclipse.jdt.ls.core.internal.managers;
 
 import java.io.File;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.eclipse.buildship.core.BuildConfiguration;
 import org.eclipse.buildship.core.GradleBuild;
@@ -25,6 +28,7 @@ import org.eclipse.buildship.core.GradleCore;
 import org.eclipse.buildship.core.internal.CorePlugin;
 import org.eclipse.buildship.core.internal.configuration.GradleProjectNature;
 import org.eclipse.buildship.core.internal.launch.GradleClasspathProvider;
+import org.eclipse.buildship.core.internal.preferences.PersistentModel;
 import org.eclipse.buildship.core.internal.util.file.FileUtils;
 import org.eclipse.buildship.core.internal.workspace.FetchStrategy;
 import org.eclipse.buildship.core.internal.workspace.InternalGradleBuild;
@@ -41,6 +45,7 @@ import org.eclipse.jdt.core.JavaCore;
 import org.eclipse.jdt.core.JavaModelException;
 import org.eclipse.jdt.ls.core.internal.JavaLanguageServerPlugin;
 import org.eclipse.jdt.ls.core.internal.ProjectUtils;
+import org.eclipse.jdt.ls.core.internal.ResourceUtils;
 import org.eclipse.jdt.ls.core.internal.managers.ProjectsManager.CHANGE_TYPE;
 import org.eclipse.jdt.ls.core.internal.preferences.IPreferencesChangeListener;
 import org.eclipse.jdt.ls.core.internal.preferences.PreferenceManager;
@@ -196,4 +201,102 @@ public class GradleBuildSupport implements IBuildSupport {
 	public boolean hasTemporaryProjectFolder() {
 		return true;
 	}
+
+	@Override
+	public void deleteInvalidProjects(Collection<IPath> rootPaths, IProgressMonitor monitor) {
+		List<String> workspaceProjects = rootPaths.stream().map((IPath rootPath) -> ProjectUtils.getWorkspaceInvisibleProjectName(rootPath)).collect(Collectors.toList());
+		List<IProject> validGradleProjects = new ArrayList<>();
+		List<IProject> suspiciousGradleProjects = new ArrayList<>();
+
+		for (IProject project : ProjectUtils.getAllProjects()) {
+			if (project.equals(ProjectsManager.getDefaultProject())) {
+				continue;
+			}
+
+			if (project.exists() && (ResourceUtils.isContainedIn(project.getLocation(), rootPaths) || applies(project) || workspaceProjects.contains(project.getName()))) {
+				try {
+					Optional<IBuildSupport> maybeBuildSupport = BuildSupportManager.find(project);
+					if (maybeBuildSupport.isPresent() && maybeBuildSupport.get().hasTemporaryProjectFolder()) {
+						project.getDescription();
+						if (ResourceUtils.isContainedIn(project.getLocation(), rootPaths)) {
+							validGradleProjects.add(project);
+						} else {
+							suspiciousGradleProjects.add(project);
+						}
+					}
+
+				} catch (CoreException e) {
+					try {
+						project.delete(true, monitor);
+					} catch (CoreException e1) {
+						JavaLanguageServerPlugin.logException(e1.getMessage(), e1);
+					}
+				}
+			} else {
+				try {
+					project.delete(false, true, monitor);
+				} catch (CoreException e1) {
+					JavaLanguageServerPlugin.logException(e1.getMessage(), e1);
+				}
+			}
+		}
+
+		List<IProject> unrelatedProjects = findUnrelatedGradleProjects(suspiciousGradleProjects, validGradleProjects);
+		unrelatedProjects.forEach((project) -> {
+			try {
+				project.delete(false, true, monitor);
+			} catch (CoreException e1) {
+				JavaLanguageServerPlugin.logException(e1.getMessage(), e1);
+			}
+		});
+	}
+
+	/**
+	 * Find those gradle projects not referenced by any gradle project in the
+	 * current workspace.
+	 */
+	private List<IProject> findUnrelatedGradleProjects(List<IProject> suspiciousProjects, List<IProject> validProjects) {
+		suspiciousProjects.sort((IProject p1, IProject p2) -> p1.getLocation().toOSString().length() - p2.getLocation().toOSString().length());
+
+		List<IProject> unrelatedCandidates = new ArrayList<>();
+		Collection<IPath> validSubPaths = new ArrayList<>();
+		for (IProject suspiciousProject : suspiciousProjects) {
+			if (validSubPaths.contains(suspiciousProject.getFullPath().makeRelative())) {
+				continue;
+			}
+
+			// Check whether the suspicious gradle project is the parent project of the opening project.
+			boolean isParentProject = false;
+			Collection<IPath> subpaths = null;
+			PersistentModel model = CorePlugin.modelPersistence().loadModel(suspiciousProject);
+			if (model.isPresent()) {
+				subpaths = model.getSubprojectPaths();
+				if (!subpaths.isEmpty()) {
+					for (IProject validProject : validProjects) {
+						if (subpaths.contains(validProject.getFullPath().makeRelative())) {
+							isParentProject = true;
+							break;
+						}
+					}
+				}
+			}
+
+			if (isParentProject) {
+				validSubPaths.addAll(subpaths);
+			} else {
+				unrelatedCandidates.add(suspiciousProject);
+			}
+		}
+
+		List<IProject> result = new ArrayList<>();
+		// Exclude those projects which are the subprojects of the verified parent project.
+		for (IProject candidate : unrelatedCandidates) {
+			if (!validSubPaths.contains(candidate.getFullPath().makeRelative())) {
+				result.add(candidate);
+			}
+		}
+
+		return result;
+	}
+
 }
