+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/managers/IBuildSupport.java
@@ -12,12 +12,14 @@
  *******************************************************************************/
 package org.eclipse.jdt.ls.core.internal.managers;
 
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResource;
 import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IPath;
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.debug.core.ILaunchConfiguration;
 import org.eclipse.jdt.core.IClassFile;
@@ -143,6 +145,9 @@ public interface IBuildSupport {
 
 	default boolean hasTemporaryProjectFolder() {
 		return false;
+	}
+
+	default void deleteInvalidProjects(Collection<IPath> rootPaths, IProgressMonitor monitor) {
 	}
 
 }
