+++ org.eclipse.jdt.ls.core/src/org/eclipse/jdt/ls/core/internal/managers/StandardProjectsManager.java
@@ -36,8 +36,6 @@ import java.util.TreeMap;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import org.eclipse.buildship.core.internal.CorePlugin;
-import org.eclipse.buildship.core.internal.preferences.PersistentModel;
 import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IFolder;
 import org.eclipse.core.resources.IProject;
@@ -118,98 +116,9 @@ public class StandardProjectsManager extends ProjectsManager {
 	}
 
 	private void deleteInvalidProjects(Collection<IPath> rootPaths, IProgressMonitor monitor) {
-		List<String> workspaceProjects = rootPaths.stream().map((IPath rootPath) -> ProjectUtils.getWorkspaceInvisibleProjectName(rootPath)).collect(Collectors.toList());
-		List<IProject> validGradleProjects = new ArrayList<>();
-		List<IProject> suspiciousGradleProjects = new ArrayList<>();
-		for (IProject project : ProjectUtils.getAllProjects()) {
-			if (project.equals(getDefaultProject())) {
-				continue;
-			}
-
-			if (project.exists() && (ResourceUtils.isContainedIn(project.getLocation(), rootPaths) ||
-									 workspaceProjects.contains(project.getName()))) {
-				try {
-					Optional<IBuildSupport> maybeBuildSupport = BuildSupportManager.find(project);
-					if (maybeBuildSupport.isPresent() && maybeBuildSupport.get().hasTemporaryProjectFolder()) {
-						project.getDescription();
-						if (ResourceUtils.isContainedIn(project.getLocation(), rootPaths)) {
-							validGradleProjects.add(project);
-						} else {
-							suspiciousGradleProjects.add(project);
-						}
-					}
-
-				} catch (CoreException e) {
-					try {
-						project.delete(true, monitor);
-					} catch (CoreException e1) {
-						JavaLanguageServerPlugin.logException(e1.getMessage(), e1);
-					}
-				}
-			} else {
-				try {
-					project.delete(false, true, monitor);
-				} catch (CoreException e1) {
-					JavaLanguageServerPlugin.logException(e1.getMessage(), e1);
-				}
-			}
+		for (IBuildSupport buildSupport : BuildSupportManager.obtainBuildSupports()) {
+			buildSupport.deleteInvalidProjects(rootPaths, monitor);
 		}
-
-		List<IProject> unrelatedProjects = findUnrelatedGradleProjects(suspiciousGradleProjects, validGradleProjects);
-		unrelatedProjects.forEach((project) -> {
-			try {
-				project.delete(false, true, monitor);
-			} catch (CoreException e1) {
-				JavaLanguageServerPlugin.logException(e1.getMessage(), e1);
-			}
-		});
-	}
-
-	/**
-	 * Find those gradle projects not referenced by any gradle project in the current workspace.
-	 */
-	private List<IProject> findUnrelatedGradleProjects(List<IProject> suspiciousProjects, List<IProject> validProjects) {
-		suspiciousProjects.sort((IProject p1, IProject p2) -> p1.getLocation().toOSString().length() - p2.getLocation().toOSString().length());
-
-		List<IProject> unrelatedCandidates = new ArrayList<>();
-		Collection<IPath> validSubPaths = new ArrayList<>();
-		for (IProject suspiciousProject : suspiciousProjects) {
-			if (validSubPaths.contains(suspiciousProject.getFullPath().makeRelative())) {
-				continue;
-			}
-
-			// Check whether the suspicious gradle project is the parent project of the opening project.
-			boolean isParentProject = false;
-			Collection<IPath> subpaths = null;
-			PersistentModel model = CorePlugin.modelPersistence().loadModel(suspiciousProject);
-			if (model.isPresent()) {
-				subpaths = model.getSubprojectPaths();
-				if (!subpaths.isEmpty()) {
-					for (IProject validProject : validProjects) {
-						if (subpaths.contains(validProject.getFullPath().makeRelative())) {
-							isParentProject = true;
-							break;
-						}
-					}
-				}
-			}
-
-			if (isParentProject) {
-				validSubPaths.addAll(subpaths);
-			} else {
-				unrelatedCandidates.add(suspiciousProject);
-			}
-		}
-
-		List<IProject> result = new ArrayList<>();
-		// Exclude those projects which are the subprojects of the verified parent project.
-		for (IProject candidate : unrelatedCandidates) {
-			if (!validSubPaths.contains(candidate.getFullPath().makeRelative())) {
-				result.add(candidate);
-			}
-		}
-
-		return result;
 	}
 
 	@Override
